requirements.txt
python-dotenv
librosa
numpy
matplotlib